+++ test/setup.ts
@@ -1,7 +1,11 @@
 import "@testing-library/jest-dom/vitest";
 import { beforeAll, afterEach, afterAll } from "vitest";
+import { cleanup } from "@testing-library/react";
 import { server } from "./mocks/server";
 
 beforeAll(() => server.listen({ onUnhandledRequest: "error" }));
-afterEach(() => server.resetHandlers());
+afterEach(() => {
+  cleanup();
+  server.resetHandlers();
+});
 afterAll(() => server.close());